@@ -1,13 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
+﻿using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
 namespace BatchDemo.Models
 {
     public class Attributes
     {
+        [JsonIgnore]
+        [IgnoreDataMember]
+        public int Id { get; set; }
         public string? Key { get; set; }
         public string? Value { get; set; }
     }
