@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using BatchDemo.Models;
+﻿using BatchDemo.Models;
 namespace BatchDemo.UnitTests
 {
     public class DemoBatchData
